more work on msvc10 build, refs #3
git-svn-id: file:///home/behr_mi/git/sumo_synched/trunk@9081 afbd958f-9f77-42d5-a016-97a22340ccf4

+++ sumo/src/microsim/MSNoLogicJunction.cpp
@@ -42,7 +42,7 @@
 // ===========================================================================
 // static member definitions
 // ===========================================================================
-std::bitset<64> MSNoLogicJunction::myDump((unsigned long) 0xffffffff);
+std::bitset<64> MSNoLogicJunction::myDump((unsigned long long) 0xffffffff);
